Refactor brush trace internals
